refactor(tests): import from TypeScript source instead of compiled output
Replaces all dist/ imports with src/ imports and adds the ts-node/esm
loader to test scripts, eliminating the need to build before running tests.

- Update all test files to import from src/ instead of dist/
- Add --no-warnings=ExperimentalWarning --loader ts-node/esm to all test scripts
- Remove preLaunchTask: build from test launch configurations
- Update telemetry test mocks from dist/.fa to src/.fa (ts-node resolves
  __dirname to src/ instead of dist/ when loading TypeScript source)
